pyproject.toml
# Build System -----------------------------------------------------------------
[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = ["src"]


# Project Metadata
[project]
name = "floodriskmap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "geopandas>=1.1.2",
    "ipykernel>=6.29.5",
    "ipython>=8.32.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "rasterio>=1.5.0",
    "xarray>=2026.2.0",
]

[project.optional-dependencies]
dev = [
    "black>=26.1.0",
    "isort>=8.0.1",
    "pandas-stubs>=2.0.0",
    "ruff>=0.15.4",
    "types-shapely>=2.0.0",
]

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]

# pyproject.toml
[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

# Ignore missing stubs for geospatial libraries
[[tool.mypy.overrides]]
module = [
    "rasterio.*",
    "geopandas.*",
]
ignore_missing_imports = true

